+++ components/database/variables.js
@@ -1,13 +1,12 @@
-import { PrismaClient } from "@prisma/client"
+import { prisma } from "../../lib/db"
 
-const prisma = new PrismaClient()
-
-async function main(query, datasetId, page, size) {
+async function main(query, include, datasetId, page, size) {
   let args = {
     select: {
       variable_id: true,
       variable_name: true,
       dataset_id: true,
+      description: true,
       dataset: {
         select: {
           dataset_name: true,
@@ -50,6 +49,44 @@ async function main(query, datasetId, page, size) {
     }
   }
 
+  const current_refreshes = [
+    "20200120",
+    "20200720",
+    "20201020",
+    "20210420",
+    "20210720",
+    "20211020",
+    "202203",
+    "202206",
+  ]
+
+  if (include !== "all") {
+    const inc = include.split("_")
+    const incs = inc.map((d) => {
+      switch (d) {
+        case "refreshes":
+          return current_refreshes
+        case "adhoc":
+          return "Adhoc"
+        case "meta":
+          return "Metadata"
+        case "rnd":
+          return "RnD"
+      }
+    })
+    const incsWhere = incs.flat().map((inc) => ({
+      refreshes: {
+        contains: inc,
+      },
+    }))
+
+    const newwhere = {
+      AND: { ...(args.where || null) },
+      OR: incsWhere,
+    }
+    args.where = newwhere
+  }
+
   const { where } = args
   const n = await prisma.variables.count({ where })
   if (n > parseInt(size)) {
@@ -67,6 +104,7 @@ async function main(query, datasetId, page, size) {
     vars: variables.map((v) => ({
       ...v,
       v_id: v.variable_id + "_" + v.dataset_id,
+      description: v.description ? true : false,
     })),
     n,
   }
@@ -74,11 +112,12 @@ async function main(query, datasetId, page, size) {
 
 export default async function getVariables(
   query = "",
+  include = "all",
   datasetId = "",
   page = 1,
   size = 10000
 ) {
-  const variables = await main(query, datasetId, page, size)
+  const variables = await main(query, include, datasetId, page, size)
     .catch((e) => {
       throw e
     })