+++ pages/_app.js
@@ -6,6 +6,7 @@ import Script from "next/script"
 import * as gtag from "../lib/gtag"
 
 import MainLayout from "../components/layout/Layout"
+import PlausibleProvider from "next-plausible"
 
 function MyApp({ Component, pageProps }) {
   const router = useRouter()
@@ -29,7 +30,13 @@ function MyApp({ Component, pageProps }) {
   const Layout = Component.Layout || MainLayout
 
   return (
-    <>
+    <PlausibleProvider
+      domain="idisearch.terourou.org"
+      customDomain="https://info.terourou.org"
+      selfHosted="true"
+      // trackLocalhost="true"
+      enabled="true"
+    >
       <Script
         strategy="afterInteractive"
         src={`https://www.googletagmanager.com/gtag/js?id=${gtag.GA_TRACKING_ID}`}
@@ -55,7 +62,7 @@ function MyApp({ Component, pageProps }) {
           setFilterTerm={setFilterTerm}
         />
       </Layout>
-    </>
+    </PlausibleProvider>
   )
 }