+++ pages/api/variables/[dataset_id]/[variable_id].js
@@ -29,18 +29,40 @@ async function main(d_id, v_id) {
               },
             },
           },
+          alternate: {
+            select: {
+              match: true,
+            },
+          },
+          matches: {
+            select: {
+              table: true,
+            },
+          },
         },
       },
       alternate: {
         select: {
-          match: true,
+          table_id: true,
+          alt_variable_id: true,
         },
       },
       matches: {
         select: {
-          variable: true,
+          table_id: true,
+          variable_id: true,
         },
       },
+      // alternate: {
+      //   select: {
+      //     variable: true,
+      //   },
+      // },
+      // matches: {
+      //   select: {
+      //     match: true,
+      //   },
+      // },
     },
   })
   const current_refreshes = [
@@ -53,6 +75,9 @@ async function main(d_id, v_id) {
     "202203",
     "202206",
   ]
+
+  if (!variable) return {}
+
   let refreshes = variable.refreshes || null
   if (refreshes && refreshes.length) {
     refreshes = current_refreshes.map((r) => ({
@@ -63,13 +88,23 @@ async function main(d_id, v_id) {
     refreshes = null
   }
 
-  const { matches, alternate, ...vble } = variable
+  const { matches, alternate, dataset, ...vble } = variable
+  const { matches: tmatch, alternate: talt, ...tbl } = variable.dataset
 
   return {
     ...vble,
+    dataset: {
+      ...tbl,
+      matches: tmatch.map((m) => m.table).concat(talt.map((a) => a.match)),
+    },
     matches: matches
-      .map((m) => m.variable)
-      .concat(alternate.map((a) => a.match)),
+      // .map((m) => m.match)
+      .concat(
+        alternate.map((a) => ({
+          table_id: a.table_id,
+          variable_id: a.alt_variable_id,
+        }))
+      ),
     database: variable.refreshes
       ? variable.refreshes.match("Adhoc|RnD|Metadata")
         ? variable.refreshes
@@ -89,5 +124,6 @@ export default async function collectionAPI(req, res) {
       await prisma.$disconnect()
     })
 
-  res.status(200).json(variable)
+  if (variable) res.status(200).json(variable)
+  res.status(400).end()
 }