+++ pages/api/variables/index.js
@@ -1,7 +1,8 @@
 import getVariables from "../../../components/database/variables"
 
 export default async function variablesAPI(req, res) {
-  const { q, datasetId, page, size } = req.query
-  const variables = await getVariables(q, datasetId, page, size)
+  res.setHeader("Cache-Control", "s-maxage=604800")
+  const { q, include, datasetId, page, size } = req.query
+  const variables = await getVariables(q, include, datasetId, page, size)
   res.status(200).json(variables)
 }